feat: rewrite everything
