+++ package.json
@@ -2,9 +2,10 @@
   "name": "discord-integration-utt",
   "version": "1.0.1",
   "description": "Discord BOT of the UTT integration",
-  "main": "index.js",
+  "main": "./src/index.ts",
   "scripts": {
-    "start": "node .",
+    "start": "ts-node ./src/index.ts",
+    "build": "tsc",
     "register": "node deploy-commands.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -14,15 +15,19 @@
   "dependencies": {
     "@discordjs/builders": "^0.15.0",
     "@discordjs/rest": "^0.5.0",
+    "@types/node": "^18.15.11",
     "axios": "^0.27.2",
     "discord-api-types": "^0.35.0",
     "discord.js": "^13.8.1",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "http-build-query": "^0.7.0",
-    "simple-json-db": "^2.0.0"
+    "simple-json-db": "^2.0.0",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.0.2"
   },
   "devDependencies": {
+    "@types/express": "^4.17.17",
     "eslint": "^8.18.0",
     "eslint-config-xo": "^0.41.0",
     "eslint-plugin-promise": "^6.0.0"
