+++ src/globals.d.ts
@@ -1,0 +1,13 @@
+declare module globalThis {
+    var data : {
+        bearer: string?,
+        bearerConfig: string?,
+        guild: any?,
+        factions: any?,
+        teams: any?,
+        rolesList: any?,
+        rolesCreatedIds: any?,
+        factionsCategoryIds: any?,
+    };
+    var db: any?;
+}
